Added ITs and some fixes
